refactor: move the common form module to the src directory
