--- packages/bezier-react/src/components/Forms/FormControl/FormControlContext.ts
@@ -1,8 +1,0 @@
-import { createContext } from '~/src/utils/react'
-
-import { type FormControlContextValue } from '~/src/components/Forms/FormControl'
-
-export const [
-  FormControlContextProvider,
-  useFormControlContext,
-] = createContext<FormControlContextValue | undefined>(undefined)
