+++ packages/bezier-react/src/components/Forms/Switch/Switch.tsx
@@ -3,7 +3,7 @@ import React, { forwardRef } from 'react'
 import * as SwitchPrimitive from '@radix-ui/react-switch'
 import classNames from 'classnames'
 
-import useFormFieldProps from '~/src/components/Forms/useFormFieldProps'
+import { useFormFieldProps } from '~/src/components/Forms/FormControl'
 
 import {
   type SwitchProps,
