--- packages/bezier-react/src/components/Forms/index.ts
@@ -1,6 +1,0 @@
-import useFormFieldProps from './useFormFieldProps'
-
-export type { FormComponentProps, FormFieldSize } from './Form.types'
-export {
-  useFormFieldProps,
-}
